requirements.txt
argparse==1.2.1
opencv-python==3.4.0.12
numpy==1.14.2
numpy-indexed==0.3.4
Pillow==5.1.0
pyflann==1.6.14
regex==2018.1.10
scipy==1.0.1
sklearn==0.0
weave==0.16.0
